Added password import, removed console logs and unnecessary comments
