+++ client/src/utils/exports.ts
@@ -78,8 +78,6 @@ export const checkVerses = (verses: Object, otherVerses: Object) => {
 export const fitVerseIn = (verses: Object, fitIn: Object, id: number) => {
   const versesKey = Object.keys(verses);
   let hold = {};
-  console.log(fitIn, id);
-  console.log(verses);
 
   delete fitIn["Part"];
   delete fitIn["Id"];
@@ -96,66 +94,8 @@ export const fitVerseIn = (verses: Object, fitIn: Object, id: number) => {
       hold[verseId + 1] = verses[verseId];
     }
   }
-
-  console.log(hold);
   return hold;
 };
-
-// export const checkExistingVerse = (
-//   verses: Object,
-//   trackNew: Object,
-//   book: string,
-//   chapter: string,
-//   verse: string,
-//   toVerse: string,
-//   versesKeys: string[],
-//   trackNewKeys: string[]
-// ) => {
-//   for (let i = 0; i < versesKeys.length; i++) {
-//     const trackId = verses[versesKeys[i]];
-//     if (
-//       trackId["Book"] === book &&
-//       trackId["Chapter"] === chapter &&
-//       trackId["Verse"] === verse &&
-//       (trackId["To"] === toVerse ||
-//         trackId["To"] === "*" ||
-//         parseInt(trackId["To"]) >= parseInt(toVerse))
-//     ) {
-//       return { status: false, message: "Verse Already Covered" };
-//     } else if (
-//       trackId["Book"] === book &&
-//       trackId["Chapter"] === chapter &&
-//       trackId["Verse"] === verse &&
-//       (trackId["To"] === "." || parseInt(trackId["To"]) > parseInt(verse)) &&
-//       toVerse === "*"
-//     ) {
-//       return { status: false, message: "* Covers A Verse That Already Exists" };
-//     }
-//   }
-//   for (let i = 0; i < trackNewKeys.length; i++) {
-//     const trackNewId = trackNew[trackNewKeys[i]];
-//     if (
-//       trackNewId["Book"] === book &&
-//       trackNewId["Chapter"] === chapter &&
-//       trackNewId["Verse"] === verse &&
-//       (trackNewId["To"] === toVerse ||
-//         trackNewId["To"] === "*" ||
-//         parseInt(trackNewId["To"]) >= parseInt(toVerse))
-//     ) {
-//       return { status: false, message: "Verse Already Covered" };
-//     } else if (
-//       trackNewId["Book"] === book &&
-//       trackNewId["Chapter"] === chapter &&
-//       trackNewId["Verse"] === verse &&
-//       (trackNewId["To"] === "." ||
-//         parseInt(trackNewId["To"]) > parseInt(verse)) &&
-//       toVerse === "*"
-//     ) {
-//       return { status: false, message: "* Covers A Verse That Already Exists" };
-//     }
-//   }
-//   return { status: true, message: "" };
-// };
 
 // OTHERS
 export const solidButton =
